combine enum files
