+++ backend/logic/order/index.js
@@ -2,12 +2,12 @@ const randomstring = require('randomstring')
 const util = require('ethereumjs-util')
 const get = require('lodash/get')
 
-const { OrderPaymentStatuses, OrderPaymentTypes } = require('../../enums')
 const { Sentry } = require('../../sentry')
 const { Order, Product } = require('../../models')
 const sendNewOrderEmail = require('../../utils/emails/newOrder')
 const discordWebhook = require('../../utils/discordWebhook')
 const { autoFulfillOrder } = require('../printful')
+const { OrderPaymentStatuses, OrderPaymentTypes } = require('../../utils/enums')
 const { decryptShopOfferData } = require('../../utils/offer')
 const { validateDiscountOnOrder } = require('../../utils/discounts')
 const { getLogger } = require('../../utils/logger')
