+++ backend/routes/crypto.js
@@ -10,7 +10,7 @@ const {
   TransactionTypes,
   TransactionStatuses,
   OrderPaymentTypes
-} = require('../enums')
+} = require('../utils/enums')
 
 const log = getLogger('routes.crypto')
