+++ backend/routes/stripe.js
@@ -15,7 +15,7 @@ const { getLogger } = require('../utils/logger')
 const makeOffer = require('./_makeOffer')
 
 const stripeWebhookErrorEmail = require('../utils/emails/stripeWebhookError')
-const { OrderPaymentTypes } = require('../enums')
+const { OrderPaymentTypes } = require('../utils/enums')
 
 const log = getLogger('routes.stripe')
