+++ backend/routes/uphold.js
@@ -8,7 +8,7 @@ const { getLogger } = require('../utils/logger')
 
 const { authShop } = require('./_auth')
 const makeOffer = require('./_makeOffer')
-const { OrderPaymentTypes } = require('../enums')
+const { OrderPaymentTypes } = require('../utils/enums')
 
 const log = getLogger('routes.uphold')
